refactor: modularize CLI command structure
- Split `packages/coursify-cli/src/commands/content.js` into resource-specific files: `courses.js`, `modules.js`, and `sections.js`.
- Move `parseFile` utility to `packages/coursify-cli/src/utils.js`.
- Update `packages/coursify-cli/src/index.js` to initialize new command groups.
- Ensure full feature parity and consistent user-facing command syntax.
