+++ packages/coursify-cli/src/index.js
@@ -4,7 +4,9 @@ import { program } from 'commander';
 import chalk from 'chalk';
 import boxen from 'boxen';
 import { setupAuthCommands } from './commands/auth.js';
-import { setupContentCommands } from './commands/content.js';
+import { setupCoursesCommands } from './commands/courses.js';
+import { setupModulesCommands } from './commands/modules.js';
+import { setupSectionsCommands } from './commands/sections.js';
 import { setupUtilsCommands } from './commands/utils.js';
 import { setupConfigCommands } from './commands/config.js';
 
@@ -26,7 +28,9 @@ program
   });
 
 setupAuthCommands(program);
-setupContentCommands(program);
+setupCoursesCommands(program);
+setupModulesCommands(program);
+setupSectionsCommands(program);
 setupUtilsCommands(program);
 setupConfigCommands(program);
